requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==24.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-cors-headers==4.3.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.1
idna==3.4
multidict==6.0.5
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
setuptools==74.0.0
sqlparse==0.4.4
twilio==9.2.4
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.4
